backend/requirements.txt
PyMuPDF
requests
pytesseract 
pillow
python-docx
num2words
pydantic
openpyxl
pandas
babel
fastapi[standard]
uvicorn[standard]
python-multipart
apscheduler
elasticsearch==8.13.0
pdf2image
pillow
requests
ollama
llama-index-llms-openai
llama-index-llms-openai-like
openai